+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/PlayerContainerFragment.kt
@@ -29,6 +29,8 @@ import au.com.shiftyjelly.pocketcasts.player.R
 import au.com.shiftyjelly.pocketcasts.player.databinding.FragmentPlayerContainerBinding
 import au.com.shiftyjelly.pocketcasts.player.view.bookmark.BookmarksFragment
 import au.com.shiftyjelly.pocketcasts.player.view.bookmark.SharedBookmarksViewModel
+import au.com.shiftyjelly.pocketcasts.player.view.chapters.ChaptersFragment
+import au.com.shiftyjelly.pocketcasts.player.view.chapters.ChaptersViewModel
 import au.com.shiftyjelly.pocketcasts.player.viewmodel.PlayerViewModel
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
 import au.com.shiftyjelly.pocketcasts.repositories.playback.UpNextSource
@@ -61,6 +63,7 @@ class PlayerContainerFragment : BaseFragment(), HasBackstack {
     private lateinit var adapter: ViewPagerAdapter
     private val viewModel: PlayerViewModel by activityViewModels()
     private val sharedBookmarksViewModel: SharedBookmarksViewModel by viewModels()
+    private val chaptersViewModel: ChaptersViewModel by activityViewModels()
     private var binding: FragmentPlayerContainerBinding? = null
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
@@ -158,7 +161,7 @@ class PlayerContainerFragment : BaseFragment(), HasBackstack {
         })
 
         viewModel.listDataLive.observe(viewLifecycleOwner) {
-            val hasChapters = it.chapters.isNotEmpty()
+            val hasChapters = !it.chapters.isEmpty
             val hasNotes = !it.podcastHeader.isUserEpisode
 
             val updated = adapter.update(hasNotes, hasChapters)
@@ -232,8 +235,9 @@ class PlayerContainerFragment : BaseFragment(), HasBackstack {
             return
         }
         binding?.viewPager?.currentItem = index
-        // HACK: View pager 2 has no way to get the actual fragments
-        childFragmentManager.fragments.filterIsInstance<ChaptersFragment>().firstOrNull()?.scrollToChapter(chapter)
+
+        // tapping on the chapter title on the now playing screen should scroll to that chapter when the fragment is available
+        chaptersViewModel.setScrollToChapter(chapter)
     }
 
     fun updateUpNextVisibility(show: Boolean) {
@@ -332,7 +336,6 @@ private class ViewPagerAdapter(fragmentManager: FragmentManager, lifecycle: Life
     }
 
     override fun createFragment(position: Int): Fragment {
-        Timber.d("Creating fragment for position $position ${sections[position]}")
         return when (sections[position]) {
             is Section.Player -> PlayerHeaderFragment()
             is Section.Notes -> NotesFragment()