+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/bookmark/BookmarksFragment.kt
@@ -100,6 +100,7 @@ class BookmarksFragment : BaseFragment() {
                     if (episodeUuid != null) {
                         BookmarksPage(
                             episodeUuid = episodeUuid,
+                            activeTheme = overrideTheme,
                             backgroundColor = requireNotNull(backgroundColor(listData)),
                             textColor = requireNotNull(textColor(listData)),
                             sourceView = sourceView,