+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/PodcastManagerImpl.kt
@@ -3,7 +3,6 @@ package au.com.shiftyjelly.pocketcasts.repositories.podcast
 import android.content.Context
 import au.com.shiftyjelly.pocketcasts.models.db.AppDatabase
 import au.com.shiftyjelly.pocketcasts.models.db.helper.TopPodcast
-import au.com.shiftyjelly.pocketcasts.models.db.helper.UserEpisodePodcastSubstitute
 import au.com.shiftyjelly.pocketcasts.models.entity.Folder
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastEpisode
@@ -41,6 +40,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
 import timber.log.Timber
 import java.util.Calendar
 import java.util.Date
@@ -681,7 +681,9 @@ class PodcastManagerImpl @Inject constructor(
 
         val uuidToAdded = HashMap<String, Boolean>()
         for (episodeUuid in episodeUuidsAdded) {
-            val episode = episodeManager.findByUuid(episodeUuid) ?: continue
+            val episode = runBlocking {
+                episodeManager.findByUuid(episodeUuid)
+            } ?: continue
             val autoDownload = podcastUuidToAutoDownload[episode.podcastUuid]
             Timber.i(
                 "Auto download " + episode.title +
@@ -744,11 +746,7 @@ class PodcastManagerImpl @Inject constructor(
     }
 
     override fun buildUserEpisodePodcast(episode: UserEpisode): Podcast {
-        return Podcast(
-            uuid = UserEpisodePodcastSubstitute.substituteUuid,
-            title = UserEpisodePodcastSubstitute.substituteTitle,
-            thumbnailUrl = episode.getUrlForArtwork()
-        )
+        return Podcast.userPodcast.copy(thumbnailUrl = episode.getUrlForArtwork())
     }
 
     override fun observeAutoAddToUpNextPodcasts(): Flowable<List<Podcast>> {